typed mouse event emitter
