+++ src/renderer/src/lib/common/EventEmitter.ts
@@ -1,49 +1,12 @@
-type EventHandler<T extends object, N extends keyof T> = (e: T[N]) => void;
+import { BaseEventEmitter } from './BaseEventEmitter';
 
 /**
  * Система обработки событий с подпиской и генерацией сигналов
  */
 export class EventEmitter<
   T extends object,
-  Name extends keyof T = keyof T,
-  Handler extends EventHandler<T, Name> = EventHandler<T, Name>
-> {
-  handlers = new Map<Name, Set<EventHandler<T, Name>>>();
-  offedOnce = new Set<Name>();
-
-  on<N extends Name>(name: N, handler: EventHandler<T, N>) {
-    if (!this.handlers.has(name)) {
-      this.handlers.set(name, new Set());
-    }
-
-    this.handlers.get(name)?.add(handler as Handler); // ? Не знаю что он тут ругается
-  }
-
-  off<N extends Name>(name: N, handler: EventHandler<T, N>) {
-    if (!this.handlers.has(name)) {
-      return;
-    }
-
-    const handlers = this.handlers.get(name);
-    handlers?.delete(handler as Handler); // ? Не знаю что он тут ругается
-
-    if (handlers?.size === 0) {
-      this.handlers.delete(name);
-    }
-  }
-
-  // Странная штука, нужна чтобы на один раз отключить событие
-  addOnceOff(name: Name) {
-    this.offedOnce.add(name);
-  }
-  removeOnceOff(name: Name) {
-    this.offedOnce.delete(name);
-  }
-
-  reset() {
-    this.handlers.clear();
-  }
-
+  Name extends keyof T = keyof T
+> extends BaseEventEmitter<T, Name> {
   emit<N extends Name>(name: N, event: T[N]) {
     if (this.offedOnce.has(name)) {
       this.offedOnce.delete(name);
