+++ src/renderer/src/lib/common/MouseEventEmitter.ts
@@ -1,3 +1,5 @@
+import { BaseEventEmitter } from './BaseEventEmitter';
+
 /**
  * Интерфейс, реализующий события, связанные с мышью.
  * Используется в {@link MouseEventEmitter}.
@@ -25,50 +27,25 @@ export enum Button {
   forward = 4,
 }
 
-/**
- * Обработчик {@link MyMouseEvent|«мышиных» событий}.
- * Используется в {@link MouseEventEmitter}.
- */
-type Handler = (e: MyMouseEvent) => any;
+interface MouseEvents {
+  mousedown: MyMouseEvent;
+  mouseup: MyMouseEvent;
+  mousemove: MyMouseEvent;
+  contextmenu: MyMouseEvent;
+  dblclick: MyMouseEvent;
+  wheel: MyMouseEvent;
+}
 
 /**
  * Система обработки событий специально для мыши.
  * Отличается от обычного {@link EventEmitter} возможностью отключить
  * вызов обработчиков событий на любом из шагов (stopPropagation).
  */
-export class MouseEventEmitter {
-  handlers = new Map<string, Set<Function>>();
-
-  addEventListener(name: string, handler: Handler) {
-    return this.on(name, handler);
-  }
-
-  on(name: string, handler: Handler) {
-    if (!this.handlers.has(name)) {
-      this.handlers.set(name, new Set());
-    }
-
-    this.handlers.get(name)?.add(handler);
-  }
-
-  off(name: string, handler: Handler) {
-    if (!this.handlers.has(name)) {
-      return;
-    }
-
-    const handlers = this.handlers.get(name);
-    handlers?.delete(handler);
-
-    if (handlers?.size === 0) {
-      this.handlers.delete(name);
-    }
-  }
-
-  reset() {
-    this.handlers.clear();
-  }
-
-  emit(name: string, event: Omit<MyMouseEvent, 'stopPropagation'>) {
+export class MouseEventEmitter<
+  T extends object = MouseEvents,
+  Name extends keyof T = keyof T
+> extends BaseEventEmitter<T, Name> {
+  emit<N extends Name>(name: N, event: Omit<T[N], 'stopPropagation'>) {
     if (this.handlers.has(name)) {
       const handlers = this.handlers.get(name);
 
@@ -84,7 +61,7 @@ export class MouseEventEmitter {
         for (let i = arr.length - 1; i >= 0; i--) {
           if (stop) break;
 
-          arr[i](event);
+          arr[i](event as T[N]);
         }
       }
     }
